pyproject.toml
[project]
name = "bsky_multitool"
version = "0.1.0"
description = "CLI + module for Bluesky data collection"
authors = [{ name="agile-engima"}]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "atproto",
  "click",
  "python-dateutil",
  "python-dotenv",
]

[project.scripts]
bsky_multitool = "bsky_multitool.__main__:cli"